+++ CHANGES.rst
@@ -106,3 +106,4 @@ v1.0.15
 v1.0.16[unreleased]
 -----------
 * Add name filtering to users
+* Deprecate /country-cities/ route and implement /available-cities/ route